requirements.txt
python-dotenv==0.19.2
Telethon==1.24.0
aiohttp==3.8.1